Add files via upload
Repository of VOG-SPEM recordings for Right Eye of Patients with and with-out Parkinson Disease with a Ground Truth and Baseline Wander delimitation.

Data were collected using a video-based eye tracking device. 52 patients and 60 controls (48 age-matched to the patients and 12 young participants) were enrolled in the study. The ground truth required to compare different techniques of BW estimation was manually delineated according to a predefined protocol. Seven methods were developed to estimate the BW, and four objective metrics were used to evaluate the results. 